+++ src/domain/entities/user.ts
