+++ src/domain/repositories/users-repository.ts
@@ -1,4 +1,4 @@
-import { User } from "../entities/user-entity";
+import { User } from "../entities/user";
 
 export interface UserRepository {
   create(data: User): Promise<User>;
