+++ src/infraestructure/database/repositories/users/users-repository.ts
@@ -1,6 +1,6 @@
 import { prisma } from "../../orm/client";
 import { UserRepository } from "../../../../domain/repositories/users-repository";
-import { User } from "../../../../domain/entities/user-entity";
+import { User } from "../../../../domain/entities/user";
 
 export class UsersRepositoryImpl implements UserRepository {
   async create(data: User): Promise<User> {
