+++ src/infraestructure/factories/controllers/user/create.ts
@@ -1,6 +1,6 @@
 import { Controller } from "../../../../domain/controller";
 import { CreateUserController } from "../../../../presentation/controllers/users/create";
-import { makeDbAddUserUseCase } from "../../usecases/user/make-db-add-user";
+import { makeDbAddUserUseCase } from "../../usecases/user/create";
 
 export const makeAddUserController = (): Controller => {
   const addUserController = new CreateUserController(makeDbAddUserUseCase());
