+++ src/presentation/routes/authentication/index.ts
@@ -1,7 +1,7 @@
 import { Router } from "express";
 import { authenticationValidatorMiddleware } from "../../validators/authentication-validator";
 import { adaptRoute } from "../../../adapters/router-adapter";
-import { makeAuthenticationController } from "../../../infraestructure/factories/controllers/authentication/make-authentication-controller";
+import { makeAuthenticationController } from "../../../infraestructure/factories/controllers/authentication/auth";
 
 export const authenticationManagementRoutes = Router();
