+++ src/presentation/routes/users/index.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import { adaptRoute } from "../../../adapters/router-adapter";
-import { makeAddUserController } from "../../../infraestructure/factories/controllers/user/make-add-controller";
+import { makeAddUserController } from "../../../infraestructure/factories/controllers/user/create";
 import { usersValidatorMiddleware } from "../../validators/users-validator";
 
 export const userManagementRoutes = Router();
