+++ src/presentation/validators/users-validator.ts
@@ -2,7 +2,7 @@ import { userValidationSchema } from "./schemas/users";
 import {
   DataValidator,
   ZodValidator,
-} from "../middlewares/zod-validator-middleware";
+} from "../middlewares/validator";
 
 export const usersValidatorMiddleware = new DataValidator(
   new ZodValidator(userValidationSchema),
